remove the extra slash at the beginning of the pathname.
this used to generate URLs like http://localhost:999//Users/foo/... which, when debugging Brackets, caused it to think it was installed at //Applications/Brackets... which caused the extension loader to ask require.js to create script tags with URLs like //Applications/Brackets... which caused Chrome to try to load http://applications/Brackets... which failed because 'applications' is not a server on my network. With only a single slash at the beginning, the extension loader generates URLs relative to the current domain once again.