don't toggle panel visibility unless necessary (fixes #48)
setPanelVisibility(), which is called many times per second as new data arrives
(or doesn't arrive), used to set the panel visibility every time, even if the
value did not change. It seems this resulted in a resize event being triggered,
which caused the CSS quick editor drop-down to disappear (see
adobe/brackets#6641).

Now the current value is read, and the value is only set if it differs from
what we want it to be.